bug fix in inline functions in helpers.hpp
